+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Add weighting to dataset search
+- Allow for searching of datasets by source tag name
 
 ## 2020-08-20